Move GIF class to the respective image frames loader
